Refactored into modules
Closes https://github.com/PistonDevelopers/quack/issues/13
