+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 
 name = "quack"
-version = "0.0.3"
+version = "0.0.4"
 authors = [
     "bvssvni <bvssvni@gmail.com",
     "Jonathan Reem <jonathan.reem@gmail.com>"
