+++ src/lib.rs
@@ -4,70 +4,14 @@
 //! A library for setting current values for stack scope,
 //! such as application structure.
 
-use std::rc::Rc;
-use std::cell::RefCell;
-use std::ops::{ Deref, DerefMut };
+pub use get::Get;
+pub use set::Set;
+pub use action::Action;
 
-impl<'a, T, U> GetFrom for (U, &'a RefCell<T>)
-    where
-        (U, T): Pair<Data = U, Object = T> + GetFrom
-{
-    #[inline(always)]
-    fn get_from(obj: &&'a RefCell<T>) -> U {
-        <(U, T) as GetFrom>::get_from(obj.borrow().deref())
-    }
-}
-
-impl<T, U> GetFrom for (U, Rc<RefCell<T>>)
-    where
-        (U, T): Pair<Data = U, Object = T> + GetFrom
-{
-    #[inline(always)]
-    fn get_from(obj: &Rc<RefCell<T>>) -> U {
-        <(U, T) as GetFrom>::get_from(obj.borrow().deref())
-    }
-}
-
-impl<'a, F, T> SetAt for (T, &'a RefCell<F>)
-    where
-        (T, F): Pair<Data = T, Object = F> + SetAt
-{
-    #[inline(always)]
-    fn set_at(val: T, obj: &mut &'a RefCell<F>) {
-        <(T, F) as SetAt>::set_at(val, obj.borrow_mut().deref_mut())
-    }
-}
-
-impl<F, T> SetAt for (T, Rc<RefCell<F>>)
-    where
-        (T, F): Pair<Data = T, Object = F> + SetAt
-{
-    #[inline(always)]
-    fn set_at(val: T, obj: &mut Rc<RefCell<F>>) {
-        <(T, F) as SetAt>::set_at(val, obj.borrow_mut().deref_mut())
-    }
-}
-
-
-impl<'a, F, A, V> ActOn<V> for (A, &'a RefCell<F>)
-    where
-        (A, F): Pair<Data = A, Object = F> + ActOn<V>
-{
-    #[inline(always)]
-    fn act_on(action: A, obj: &mut &'a RefCell<F>) -> V {
-        <(A, F) as ActOn<V>>::act_on(action, obj.borrow_mut().deref_mut())
-    }
-}
-
-impl<F, A, V> ActOn<V> for (A, Rc<RefCell<F>>)
-    where
-        (A, F): Pair<Data = A, Object = F> + ActOn<V>
-{
-    #[inline(always)]
-    fn act_on(action: A, obj: &mut Rc<RefCell<F>>) -> V {
-        <(A, F) as ActOn<V>>::act_on(action, obj.borrow_mut().deref_mut())
-    }
-}
+mod refcell;
+mod get;
+mod set;
+mod action;
 
 /// Something that can be set at an object.
 ///
@@ -81,33 +25,6 @@ pub trait SetAt: Pair {
     );
 }
 
-/// Automatically implemented through the `SetAt` trait.
-#[unstable]
-pub trait Set<T> {
-    /// Set value.
-    fn set(mut self, val: T) -> Self;
-
-    /// Set value through mutable reference.
-    fn set_mut(&mut self, val: T) -> &mut Self;
-}
-
-impl<T, U> Set<U> for T
-    where
-        (U, T): Pair<Data = U, Object = T> + SetAt,
-{
-    #[inline(always)]
-    fn set(mut self, val: U) -> T {
-        <(U, T) as SetAt>::set_at(val, &mut self);
-        self
-    }
-
-    #[inline(always)]
-    fn set_mut(&mut self, val: U) -> &mut T {
-        <(U, T) as SetAt>::set_at(val, self);
-        self
-    }
-}
-
 /// Something that can be retrieved from another object.
 #[unstable]
 pub trait GetFrom: Pair {
@@ -115,23 +32,6 @@ pub trait GetFrom: Pair {
     fn get_from(
         obj: &<Self as Pair>::Object
     ) -> <Self as Pair>::Data;
-}
-
-/// Automatically implemented through the `GetFrom` trait.
-#[unstable]
-pub trait Get<T> {
-    /// Returns new value.
-    fn get(&self) -> T;
-}
-
-impl<T, U> Get<U> for T
-    where
-        (U, T): Pair<Data = U, Object = T> + GetFrom
-{
-    #[inline(always)]
-    fn get(&self) -> U {
-        <(U, T) as GetFrom>::get_from(self)
-    }
 }
 
 /// Does something to an object.
@@ -144,23 +44,6 @@ pub trait ActOn<T>: Pair {
     ) -> T;
 }
 
-/// Automatically implemented through the `ActOn` trait.
-#[unstable]
-pub trait Action<A, V> {
-    /// Does something.
-    fn action(&mut self, val: A) -> V;
-}
-
-impl<T, A, V> Action<A, V> for T
-    where
-        (A, T): Pair<Data = A, Object = T> + ActOn<V>
-{
-    #[inline(always)]
-    fn action(&mut self, action: A) -> V {
-        <(A, T) as ActOn<V>>::act_on(action, self)
-    }
-}
-
 /// Used to reduce the need for associated types.
 pub trait Pair {
     type Data;
@@ -171,4 +54,3 @@ impl<T, U> Pair for (T, U) {
     type Data = T;
     type Object = U;
 }
-
